requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
dataclasses-json==0.6.3
decorator==4.4.2
exceptiongroup==1.2.0
ffmpeg-python==0.2.0
filelock==3.13.1
frozenlist==1.4.1
fsspec==2023.12.2
future==0.18.3
greenlet==3.0.2
idna==3.6
imageio==2.33.1
imageio-ffmpeg==0.4.9
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.352
langchain-community==0.0.5
langchain-core==0.1.2
langsmith==0.0.72
llvmlite==0.41.1
MarkupSafe==2.1.3
marshmallow==3.20.1
more-itertools==10.1.0
moviepy==1.0.3
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.2.1
numba==0.58.1
numpy==1.26.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
openai-whisper==20231117
packaging==23.2
Pillow==10.1.0
proglog==0.1.10
pydantic==2.5.2
pydantic_core==2.14.5
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==2.0.23
sympy==1.12
tenacity==8.2.3
tiktoken==0.5.2
torch==2.1.2
torchaudio==2.1.2
torchvision==0.16.2
tqdm==4.66.1
triton==2.1.0
typing-inspect==0.9.0
typing_extensions==4.9.0
urllib3==2.1.0
yarl==1.9.4
